+++ examples/autoencoder_example.py
@@ -1,10 +1,10 @@
-from sapsan.general.backends.fake import FakeExperimentBackend
-from sapsan.general.data.jhtdb_dataset import JHTDB128Dataset
-from sapsan.general.data.sampling.equidistant_sampler import Equidistance3dSampling
-from sapsan.general.estimator.cnn.spacial_autoencoder import SpacialAutoencoderNetworkEstimator,\
+from sapsan.core.backends.fake import FakeExperimentBackend
+from sapsan.core.data.jhtdb_dataset import JHTDB128Dataset
+from sapsan.core.data.sampling.equidistant_sampler import Equidistance3dSampling
+from sapsan.core.estimator.cnn.spacial_autoencoder import SpacialAutoencoderNetworkEstimator,\
     SpacialAutoencoderNetworkEstimatorConfiguration
-from sapsan.general.experiments.evaluation_autoencoder import EvaluationAutoencoderExperiment
-from sapsan.general.experiments.training import TrainingExperiment
+from sapsan.core.experiments.evaluation_autoencoder import EvaluationAutoencoderExperiment
+from sapsan.core.experiments.training import TrainingExperiment
 
 
 def run():
