+++ kubeflow/pipelines/mock_pipeline/training/train.py
@@ -4,9 +4,9 @@ import sys
 import time
 import json
 
-from sapsan.general.mocks.fake_dataset import FakeDataset
-from sapsan.general.mocks.regression_estimator import LinearRegressionEstimator, LinearRegressionEstimatorConfiguration
-from sapsan.general.mocks.fake_experiment import FakeTrainingExperiment, FakeInferenceExperiment
+from sapsan.core.mocks.fake_dataset import FakeDataset
+from sapsan.core.mocks.regression_estimator import LinearRegressionEstimator, LinearRegressionEstimatorConfiguration
+from sapsan.core.mocks.fake_experiment import FakeTrainingExperiment, FakeInferenceExperiment
 
 
 logging.basicConfig(level=logging.INFO,
