+++ sapsan/core/estimator/cnn/spacial_3d_encoder.py
@@ -22,9 +22,9 @@ from typing import Dict
 import torch
 from catalyst.dl import SupervisedRunner, EarlyStoppingCallback
 
-from sapsan.general.data.jhtdb_dataset import JHTDBDatasetPyTorchSplitterPlugin, JHTDB128Dataset, \
+from sapsan.core.data.jhtdb_dataset import JHTDBDatasetPyTorchSplitterPlugin, JHTDB128Dataset, \
     OutputFlatterDatasetPlugin
-from sapsan.general.models import Estimator, EstimatorConfiguration
+from sapsan.core.models import Estimator, EstimatorConfiguration
 
 
 class Spacial3dEncoderNetworkModel(torch.nn.Module):
